requirements.txt
mock==4.0.3
tensorflow_gpu==2.5.0
scikit_image==0.18.1
numpy==1.20.2
